requirements.txt
# DogTrackerV2/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-WTF>=1.0
python-dotenv>=0.19
SQLAlchemy>=1.4
alembic>=1.7
# Add waitress or gunicorn later if needed